Quicksave

+++ BandView/views.py
@@ -83,34 +83,34 @@ def signup(request):
 
 
 
-def bandsignup(request):
-    if request.method == 'POST':
-        form = BandForm(request.POST or None)
-        if form.is_valid():
-            band = form.save(commit=False)
-            band.user = request.user
-            form.save()
-            return redirect('bands')
-        else:
-            return HttpResponse("Form is not valid")
-    else:
-        form = BandForm
-        return render(request, 'BandSignUp.html', {'form': form})
-
-
-def venuesignup(request):
-    if request.method == 'POST':
-        form = VenueForm(request.POST or None)
-        if form.is_valid():
-            venue = form.save(commit=False)
-            venue.user = request.user
-            form.save()
-            return redirect('venues')
-        else:
-            return HttpResponse("Form is not valid")
-    else:
-        form = VenueForm
-        return render(request, 'AddVenue.html', {'form': form})
+# def bandsignup(request):
+#     if request.method == 'POST':
+#         form = BandForm(request.POST or None)
+#         if form.is_valid():
+#             band = form.save(commit=False)
+#             band.user = request.user
+#             form.save()
+#             return redirect('bands')
+#         else:
+#             return HttpResponse("Form is not valid")
+#     else:
+#         form = BandForm
+#         return render(request, 'BandSignUp.html', {'form': form})
+#
+#
+# def venuesignup(request):
+#     if request.method == 'POST':
+#         form = VenueForm(request.POST or None)
+#         if form.is_valid():
+#             venue = form.save(commit=False)
+#             venue.user = request.user
+#             form.save()
+#             return redirect('venues')
+#         else:
+#             return HttpResponse("Form is not valid")
+#     else:
+#         form = VenueForm
+#         return render(request, 'AddVenue.html', {'form': form})
 
 
 def bands(request):
@@ -145,45 +145,69 @@ def venueprofile(request, venue_id):
 
 
 def updateBand(request, band_id):
-    band = Band.objects.get(pk=band_id)
-    form = BandForm(request.POST or None, instance=band)
-    if request.method == 'POST':
-        if form.is_valid():
-            form.save()
-            return redirect('bands')
-    context = {'form': form}
-    return render(request, 'updateBand.html', context)
+    if request.user.is_authenticated:
+        band = Band.objects.get(pk=band_id)
+        if request.user.id == band.user.id:
+            form = BandForm(request.POST or None, instance=band)
+            if request.method == 'POST':
+                if form.is_valid():
+                    form.save()
+                    return redirect('bands')
+            context = {'form': form}
+            return render(request, 'updateBand.html', context)
+        else:
+            return HttpResponse("Permission denied")
+    else:
+        redirect('notauthenticated')
 
 
 def deleteBand(request, band_id):
-    band = Band.objects.get(pk=band_id)
-    if request.method == 'POST':
-        band.delete()
-        return redirect('bands')
+    if request.user.is_authenticated:
+        band = Band.objects.get(pk=band_id)
+        if request.user.id == band.user.id:
+            if request.method == 'POST':
+                band.delete()
+                return redirect('bands')
+            else:
+                context = {'band': band}
+                return render(request, 'deleteBand.html', context)
+        else:
+            return HttpResponse("Permission denied")
     else:
-        context = {'band': band}
-        return render(request, 'deleteBand.html', context)
+        redirect('notauthenticated')
 
 
 def updateVenue(request, venue_id):
-    venue = Venue.objects.get(pk=venue_id)
-    form = VenueForm(request.POST or None, instance=venue)
-    if request.method == 'POST':
-        if form.is_valid():
-            form.save()
-            return redirect('venues')
-    context = {'form': form}
-    return render(request, 'updateVenue.html', context)
+    if request.user.is_authenticated:
+        venue = Venue.objects.get(pk=venue_id)
+        if request.user.id == venue.user.id:
+            form = VenueForm(request.POST or None, instance=venue)
+            if request.method == 'POST':
+                if form.is_valid():
+                    form.save()
+                    return redirect('venues')
+            context = {'form': form}
+            return render(request, 'updateVenue.html', context)
+        else:
+            return HttpResponse("Permission denied")
+    else:
+        redirect('notauthenticated')
 
 
 def deleteVenue(request, venue_id):
-    venue = Band.objects.get(pk=venue_id)
-    if request.method == 'POST':
-        venue.delete()
-        return redirect('venues')
+    if request.user.is_authenticated:
+        venue = Band.objects.get(pk=venue_id)
+        if request.user.id == venue.user.id:
+            if request.method == 'POST':
+                venue.delete()
+                return redirect('venues')
+            else:
+                context = {'venue': venue}
+                return render(request, 'deleteVenue.html', context)
+        else:
+            return HttpResponse("Permission denied")
     else:
-        context = {'venue': venue}
-        return render(request, 'deleteVenue.html', context)
+        redirect('notauthenticated')
 
 def createEvent(request):
     if request.method == 'POST':
@@ -202,3 +226,6 @@ def createEvent(request):
 def signout(request):
     logout(request)
     return redirect('welcome')
+
+def not_authenticated(request):
+    return render(request, 'not_authenticated.html')
